Wip - found a bug in ProductSubscriber

+++ ReadService/Data/Models/Product.cs
@@ -5,7 +5,7 @@ namespace ReadService.Data.Models;
 public sealed record Product
 {
     [BsonId]
-    public int Id { get; set; }
+    public long Id { get; set; }
 
     [BsonElement("title")]
     public required string Title { get; set; }
@@ -20,7 +20,7 @@ public sealed record Product
     public required decimal Price { get; set; }
 
     [BsonElement("rating")]
-    public required int Rating { get; set; }
+    public required double Rating { get; set; }
 
     [BsonElement("vendor")]
     public required ProductVendor Vendor { get; set; }
@@ -35,7 +35,7 @@ public sealed record Product
 public sealed record ProductVendor
 {
     [BsonId]
-    public int Id { get; set; }
+    public long Id { get; set; }
 
     [BsonElement("name")]
     public required string Name { get; set; }
