+++ ReadService/Data/Models/ProductsOfCategory.cs
@@ -13,7 +13,7 @@ public sealed record ProductsOfCategory
     public sealed record Product
     {
         [BsonId]
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         [BsonElement("title")]
         public required string Title { get; set; }
@@ -25,7 +25,7 @@ public sealed record ProductsOfCategory
         public required decimal Price { get; set; }
 
         [BsonElement("rating")]
-        public required int Rating { get; set; }
+        public required double Rating { get; set; }
 
         [BsonElement("vendor")]
         public required ProductVendor Vendor { get; set; }
@@ -33,7 +33,7 @@ public sealed record ProductsOfCategory
         public sealed record ProductVendor
         {
             [BsonElement("_id")]
-            public int Id { get; set; }
+            public long Id { get; set; }
 
             [BsonElement("name")]
             public required string Name { get; set; }
